+++ frontend/src/components/Admin/User/UserEditForm.jsx
@@ -1,12 +1,11 @@
 import { useEffect, useState } from "react"
 import { useNavigate, useParams } from "react-router-dom"
 
-const UsersEditForm = ({ user }) => {
+const UserEditForm = () => {
 
     const { id } = useParams()
     const [userToEdit, setUserToEdit] = useState([])
     const navigate = useNavigate()
-
 
     const getUser = () => {
         fetch('/api/users/' + id)
@@ -24,7 +23,6 @@ const UsersEditForm = ({ user }) => {
         getUser()
     }, [])
 
-
     const handleEdit = (e) => {
         e.preventDefault()
         const elements = e.target.elements
@@ -41,37 +39,40 @@ const UsersEditForm = ({ user }) => {
         })
             .then(response => {
                 if (!response.ok) throw new Error('Network response')
-                navigate('/admin/users');
+                navigate('/admin/users')
             })
             .catch(err => console.log('Error during fetch', err))
     }
 
     const handleDelete = (userId) => {
-        fetch('/api/users/' + userId, { method: 'DELETE' })
-            .then(response => {
-                if (!response.ok) throw new Error('Nepavyko delete')
-                navigate('/admin/users');
-            })
-            .catch(err => console.log('Error during delete', err))
+        const confirm = window.confirm("Ar you sure you want to DELETE this order?")
+        if (confirm) {
+            fetch('/api/users/' + userId, { method: 'DELETE' })
+                .then(response => {
+                    if (!response.ok) throw new Error('Nepavyko delete')
+                    navigate('/admin/users')
+                })
+                .catch(err => console.log('Error during delete', err))
+        }
     }
 
     return (
         <div className="form-container">
-                <div className="form-styles">
-            <h2>Edit Users</h2>
-            <form onSubmit={handleEdit}>
-                <label htmlFor="email">Email</label>
-                <input type="text" name="email" id="email" defaultValue={userToEdit.email} required />
-                <label htmlFor="name">Name</label>
-                <input type="text" name="name" id="name" defaultValue={userToEdit.name} required />
-                <label htmlFor="surname">Surname</label>
-                <input type="text" name="surname" id="surname" defaultValue={userToEdit.surname} required />
-                <button type="submit" className="button-styles">Update</button>
-                <button onClick={() => handleDelete(userToEdit._id)} className="button-styles">Delete</button>
-            </form>
-        </div>
+            <div className="form-styles">
+                <h2>Edit Users</h2>
+                <form onSubmit={handleEdit}>
+                    <label htmlFor="email">Email</label>
+                    <input type="text" name="email" id="email" defaultValue={userToEdit.email} required />
+                    <label htmlFor="name">Name</label>
+                    <input type="text" name="name" id="name" defaultValue={userToEdit.name} required />
+                    <label htmlFor="surname">Surname</label>
+                    <input type="text" name="surname" id="surname" defaultValue={userToEdit.surname} required />
+                    <button type="submit" className="button-styles">Update</button>
+                    <button type="button" onClick={() => handleDelete(userToEdit._id)} className="button-styles">Delete</button>
+                </form>
+            </div>
         </div>
     );
 }
 
-export default UsersEditForm;
+export default UserEditForm;
